+++ src/py/pyodide/http.py
@@ -13,6 +13,12 @@ from ._core import IN_BROWSER
 if IN_BROWSER:
     from js import fetch as _jsfetch, Object
 
+__all__ = [
+    "open_url",
+    "pyfetch",
+    "FetchResponse",
+]
+
 
 def open_url(url: str) -> StringIO:
     """Fetches a given URL synchronously.
@@ -156,7 +162,7 @@ async def pyfetch(url, **kwargs) -> FetchResponse:
 
     Parameters
     ----------
-    url URL to fetch. **kwargs Any keyword arguments are passed along as
+    url URL to fetch. \*\*kwargs Any keyword arguments are passed along as
         `optional parameters to the fetch API
         <https://developer.mozilla.org/en-US/docs/Web/API/fetch#parameters>`_.
     """